requirements.txt
blinker==1.9.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lightning-utilities==0.12.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.6.1
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
